+++ CHANGELOG.md
@@ -8,6 +8,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ### Changed
 
 ### Fixed
+- Show a user-facing error when moving a feed fails (#3649)
 
 # Releases
 ## [28.2.0-beta.1] - 2026-03-22